reviews_api/requirements/pip-base.txt
Django==2.0.4
dj-database-url==0.4.2
djangorestframework-jwt==1.11.0
djangorestframework==3.8.0
django-environ==0.4.4
psycopg2==2.7.3.1
pytz==2018.3
raven==6.6.0
whitenoise==3.3.1
